Disable cron for new database during initialisation
During database creation CRON error can be observed in the error log.
In slower machines the timing of the errors can cause a deadlock state as
the database is added to the cron while the cron tables are being initialised.
This leaves the database in an inconsistent and unusable state for the user.

To correct this I have exposed the pool lock by adding a lock property, and
required the server to acquire a lock on the database pool before it can start a
thread for it.

issue2234
review173001