Use value instead of record instance on import_data
Importing records with m2o- and m2m-relations import_data gives record instances
to the create-method of the object instead of record-id's. Creating the same
object using the GUI the client gives record-id's.
On error the record-instances can't be serialized by JSON and the common error
"500 internal server error" will be reported to the client

issue3124
review767002