Reset MPTT fields when copying records
The left/right fields must be reset to their default values to allow to compute
the correct new values. As it is common to forget this requirement when
implementing, it is safer to enforce this behaviour directly in the
ModelStorage.copy method.

issue5049
review16741002
(grafted from f1997c1aa66c123c21402508a5c014d9c915a442)

--HG--
branch : 2.8