Fix default value and check of last_timestamp of Sequence
The client integer widget does not allow to display float value. So the default
value of last_timestamp is replaced by None which fails later in the check.
Also the check must not validate when it is not filled because in Python3 None
and int are unorderable.

issue5591
review28571002
(grafted from c513c0893d086b381dd55c71f2a26005cd2a1d4d)

--HG--
branch : 3.2