Ensure Many2One target join is unique if Target has history
The joins added in tables must ensure to make a Cartesian product.
So when joining on a history table, the target id is not unique so the latest
must be used only for the join. Otherwise a search query may return multiple
times the same id.

issue6458
review37341002
(grafted from 1326d29d969c5fa41f723e5c32bcaaaaba5e5382)

--HG--
branch : 3.8