Remove useless computation of module on the translations creation
The .create() call for the ir.translation object guess the module of the
translations created when the module passed is not a value evaluating to True.

This could cause the unique constraint to fail when synchronizing translations
after a report has been overriden in the interface.

issue6732
review36581002
(grafted from 17d1db4a31a9643965503142e99208e5c22b43ef)

--HG--
branch : 4.0